pyproject.toml
[tool.poetry]
name = "deepsafe"
version = "0.1.0"
description = "Open source deepfake detection"
authors = ["Siddharth Kumar <siddharth123sk@gmail.com>"]
license = "Dual Licenses"

[tool.poetry.dependencies]
python = "^3.9"
streamlit = "1.24.0"
numpy = "1.24.3"
opencv-python-headless = "4.7.0.72"
pandas = "1.5.3"
Pillow = "9.5.0"
scikit-learn = "1.2.2"
scikit-image = "0.20.0"
tqdm = "4.65.0"
tensorflow = "2.12.0"
torch = "2.0.0"
torchvision = "0.15.0"
matplotlib = "3.7.1"
altair = "4.1.0"
protobuf = "3.20.*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"